Hardness should be a float

+++ src/main/java/org/waveapi/api/content/items/block/WaveBlock.java
@@ -26,6 +26,6 @@ public class WaveBlock {
 
     public WaveBlock setTab(WaveTab tab) {return this;}
 
-    public WaveBlock setHardness(int hardness) {return this; }
+    public WaveBlock setHardness(float hardness) {return this; }
 
 }
